requirements.txt
tqdm==4.32.1
gpflow==1.4.1
flask==1.0.2
tensorboardX==1.8
absl-py==0.6.1
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
boto3==1.9.57
botocore==1.12.57
cachetools==3.0.0
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.3.9
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.1
dask==1.0.0
decorator==4.3.0
docutils==0.14
dotmap==1.3.8
deepdiff==3.3.0
flatbuffers==1.10
funcsigs==1.0.2
future==0.17.1
gast==0.2.0
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.7.0
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
grpcio==1.16.1
gtimer==1.0.0b5
gym==0.12.0
h5py==2.8.0
httplib2==0.12.0
idna==2.7
imageio==2.4.1
jmespath==0.9.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.0.1
matplotlib==3.0.2
more-itertools==4.3.0
git+https://github.com/vitchyr/multiworld.git@d76b3dae2e8cbca02924f93d6cc0239c552f6408
networkx==2.2
numpy==1.15.4
pandas==0.23.4
plotly==1.9.6
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.19
pygame==1.9.4
pyglet==1.3.2
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.0.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==4.2b4
psutil==5.4.8
ray[rllib,debug]==0.6.4
redis==3.0.1
requests==2.20.1
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.2
scikit-learn==0.20.1
scipy==1.1.0
git+https://github.com/hartikainen/serializable.git@76516385a3a716ed4a2a9ad877e2d5cbcf18d4e6
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.5
tensorboard==1.13.1
tensorflow-gpu==1.13.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.6.0
termcolor==1.1.0
toolz==0.9.0
uritemplate==3.0.0
urllib3==1.24.2
Werkzeug==0.15.3
torch==1.9.0
meshcat==0.3.1